Moved database tests to folder
