+++ tests/test_database/test_common.py
@@ -9,8 +9,8 @@ from OpenPostbud.database.common import ShipmentStatus
 from OpenPostbud.database.digital_post import letters
 from OpenPostbud.database.digital_post.letters import Letter
 from OpenPostbud.database.digital_post.shipments import Shipment
-from tests import builders
-from tests.db_test_case import DBTestCase
+from tests.test_database import builders
+from tests.test_database.db_test_case import DBTestCase
 
 
 class ShipmentStatusEnumTest(unittest.TestCase):
