+++ tests/test_database/test_encrypted_string.py
@@ -5,8 +5,8 @@ import unittest
 from sqlalchemy import text
 
 from OpenPostbud.database.digital_post import letters
-from tests import builders
-from tests.db_test_case import DBTestCase
+from tests.test_database import builders
+from tests.test_database.db_test_case import DBTestCase
 
 
 class EncryptedStringTest(DBTestCase):
